+++ src/components/Comment.vue
@@ -5,27 +5,25 @@
     <SubTitle :title="'titles.comment'" />
     <div id="gitalk-container"></div>
     <div id="vcomments"></div>
+    <div id="tcomment"></div>
+    <div id="waline"></div>
   </div>
 </template>
 
 <script lang="ts">
-/**
- * Gitalk and Valine is imported through CDN
- * need to dereferencing these library globals
- *
- * Gitalk version 1.6
- * Valine version 1.4.14
- *
- * Check the usage of these version before
- * using these global variables.
- */
-declare const Gitalk: any
-declare const Valine: any
-
 import { useAppStore } from '@/stores/app'
 import { defineComponent, onMounted, toRefs, watch } from 'vue'
 import { SubTitle } from '@/components/Title'
 import { usePostStore } from '@/stores/post'
+import { twikooInit } from '@/utils/comments/twikoo-api'
+import { githubInit } from '@/utils/comments/github-api'
+import { valineInit } from '@/utils/comments/valine-api'
+import { walineInit } from '@/utils/comments/waline-api'
+
+const languages: Record<string, string> = {
+  en: 'en',
+  cn: 'zh'
+}
 
 export default defineComponent({
   name: 'ObComment',
@@ -53,6 +51,7 @@ export default defineComponent({
     const postUid = toRefs(props).uid
     const appStore = useAppStore()
     const postStore = usePostStore()
+    let waline: any = undefined
 
     const enabledComment = (
       postTitle: string,
@@ -91,35 +90,73 @@ export default defineComponent({
             ? 'https://cors-anywhere.azm.workers.dev/https://github.com/login/oauth/access_token'
             : appStore.themeConfig.plugins.gitalk.proxy
 
-        const gitalk = new Gitalk({
-          clientID: appStore.themeConfig.plugins.gitalk.clientID,
-          clientSecret: appStore.themeConfig.plugins.gitalk.clientSecret,
-          repo: appStore.themeConfig.plugins.gitalk.repo, // The repository of store comments,
-          owner: appStore.themeConfig.plugins.gitalk.owner,
-          admin: appStore.themeConfig.plugins.gitalk.admin,
-          id: uid, // Ensure uniqueness and length less than 50
-          language: appStore.themeConfig.plugins.gitalk.language,
-          distractionFreeMode: true, // Facebook-like distraction free mode
-          title: title,
-          body: body,
-          proxy: proxy
-        })
+        const { clientID, clientSecret, repo, owner, admin, language } =
+          appStore.themeConfig.plugins.gitalk
 
-        gitalk.render('gitalk-container')
+        githubInit({
+          clientID,
+          clientSecret,
+          repo,
+          owner,
+          admin,
+          language,
+          uid,
+          title,
+          body,
+          proxy
+        })
       } else if (appStore.themeConfig.plugins.valine.enable) {
-        new Valine({
-          el: '#vcomments',
-          appId: appStore.themeConfig.plugins.valine.app_id,
-          appKey: appStore.themeConfig.plugins.valine.app_key,
-          avatar: appStore.themeConfig.plugins.valine.avatar,
-          placeholder: appStore.themeConfig.plugins.valine.placeholder,
-          visitor: appStore.themeConfig.plugins.valine.visitor,
-          lang: appStore.themeConfig.plugins.valine.lang,
-          meta: appStore.themeConfig.plugins.valine.meta ?? [],
-          requiredFields:
-            appStore.themeConfig.plugins.valine.requiredFields ?? [],
-          avatarForce: appStore.themeConfig.plugins.valine.avatarForce,
+        const {
+          app_id,
+          app_key,
+          avatar,
+          placeholder,
+          visitor,
+          lang,
+          meta,
+          requiredFields,
+          avatarForce
+        } = appStore.themeConfig.plugins.valine
+
+        valineInit({
+          appId: app_id,
+          appKey: app_key,
+          avatar,
+          placeholder,
+          visitor,
+          lang,
+          meta,
+          requiredFields,
+          avatarForce,
           path: window.location.pathname // Make sure updating pathname
+        })
+      } else if (appStore.themeConfig.plugins.twikoo.enable) {
+        const { envId, region, lang } = appStore.themeConfig.plugins.twikoo
+        twikooInit({ envId, region, lang, path: window.location.pathname })
+      } else if (appStore.themeConfig.plugins.waline.enable) {
+        const {
+          serverURL,
+          login,
+          reaction,
+          meta,
+          requiredMeta,
+          commentSorting,
+          wordLimit,
+          imageUploader,
+          pageSize
+        } = appStore.themeConfig.plugins.waline
+
+        waline = walineInit({
+          serverURL,
+          lang: languages[appStore.locale ?? 'en'],
+          login,
+          reaction,
+          meta,
+          requiredMeta,
+          commentSorting,
+          wordLimit,
+          imageUploader,
+          pageSize
         })
       }
     }
@@ -131,6 +168,19 @@ export default defineComponent({
         if (!oldValue && newValue) {
           const cachePost = postStore.cachePost
           enabledComment(cachePost.title, cachePost.body, cachePost.uid)
+        }
+      }
+    )
+
+    /** Updating comments base on current locale */
+    watch(
+      () => appStore.locale,
+      (newLocale, oldLocale) => {
+        console.log(newLocale, oldLocale)
+        if (waline && newLocale !== undefined && newLocale !== oldLocale) {
+          waline.update({
+            lang: languages[newLocale]
+          })
         }
       }
     )
@@ -326,6 +376,120 @@ export default defineComponent({
         }
       }
     }
+  }
+}
+
+#twikoo {
+  .tk-comments {
+    @apply pt-2;
+  }
+
+  .tk-input {
+    @apply bg-ob-deep-900;
+  }
+
+  .tk-meta-input input {
+    @apply bg-ob-deep-900;
+  }
+
+  .el-input__inner:focus,
+  .el-textarea__inner:focus {
+    border-color: var(--text-accent);
+  }
+
+  .el-button--primary {
+    background: var(--main-gradient);
+    border: none;
+    color: #fff;
+    &:hover {
+      color: #fff;
+      opacity: 0.5;
+    }
+  }
+
+  .tk-icon.__comments {
+    color: var(--text-accent);
+  }
+
+  .tk-action-icon {
+    color: var(--text-accent);
+  }
+
+  .tk-comment {
+    @apply px-4 pt-6 pb-6 bg-ob-deep-900 mb-2 rounded-lg;
+    transition: var(--trans-ease);
+    &:hover {
+      box-shadow: var(--accent-shadow);
+    }
+  }
+
+  .tk-avatar {
+    border: none;
+  }
+
+  .tk-nick {
+    color: var(--text-accent);
+    font-weight: 700;
+  }
+
+  .tk-comments-count {
+    > span:first-of-type {
+      color: var(--text-accent);
+    }
+  }
+}
+
+#waline {
+  --waline-theme-color: var(--text-accent);
+  --waline-border: var(--background-secondary);
+  --waline-bgcolor: var(--background-primary);
+  --waline-bgcolor-light: var(--background-secondary);
+  --waline-badge-color: var(--text-accent);
+  --waline-disabled-bgcolor: var(--text-dim);
+  .wl-editor {
+    @apply p-2 box-border;
+  }
+
+  .wl-login-nick,
+  .wl-nick {
+    color: var(--text-sub-accent);
+  }
+
+  .wl-card {
+    @apply bg-ob-deep-900 p-4 rounded-lg;
+  }
+
+  .wl-num {
+    color: var(--text-accent);
+  }
+
+  .primary.wl-btn {
+    color: var(--text-bright);
+    border: none;
+    background: var(--main-gradient);
+    transition: var(--trans-ease);
+    &:hover {
+      opacity: 0.5;
+    }
+  }
+
+  .wl-card .wl-delete,
+  .wl-card .wl-like,
+  .wl-card .wl-reply,
+  .wl-card .wl-edit {
+    color: var(--text-dim);
+  }
+
+  .wl-card .wl-quote {
+    border-inline-start: none;
+  }
+
+  .wl-header-item {
+    @apply items-center;
+  }
+
+  .wl-header input {
+    @apply p-2 m-2;
   }
 }
 </style>