+++ src/components/Sidebar/src/RecentComment.vue
@@ -1,22 +1,25 @@
 <template>
-  <div class="sidebar-box">
+  <div v-if="!!enabledCommentPlugin" class="sidebar-box">
     <SubTitle :title="'titles.recent_comment'" icon="quote" />
     <ul>
-      <template v-if="comments.length > 0">
+      <template v-if="isLoading === false">
         <li
-          class="bg-ob-deep-900 px-2 py-3 mb-1.5 rounded-lg flex flex-row justify-items-center items-center shadow-sm hover:shadow-ob transition-shadow"
+          class="bg-ob-deep-900 px-2 py-2 mb-1.5 rounded-lg flex flex-row justify-items-center items-stretch shadow-sm hover:shadow-ob transition-shadow"
+          v-if="comments.length > 0"
           v-for="comment in comments"
           :key="comment.id"
         >
-          <img
-            class="col-span-1 mr-2 rounded-full p-1"
-            :src="comment.user.avatar_url"
-            alt="comment-avatar"
-            height="40"
-            width="40"
-          />
+          <div class="flex justify-start items-start">
+            <img
+              class="col-span-1 mr-2 rounded-full p-1"
+              :src="comment.user.avatar_url"
+              alt="comment-avatar"
+              height="40"
+              width="40"
+            />
+          </div>
           <div class="flex-1 text-xs">
-            <div class="text-xs">
+            <div class="text-xs mb-2 pt-1">
               <span class="text-ob pr-2">
                 {{ comment.user.login }}
                 <b
@@ -28,11 +31,24 @@
               </span>
               <p class="text-gray-500">{{ comment.created_at }}</p>
             </div>
-            <div class="text-xs text-ob-bright">
+            <div class="text-xs text-ob-bright pb-1">
               {{ comment.body }}
             </div>
           </div>
         </li>
+
+        <div
+          v-else
+          class="flex flex-row justify-center items-center text-ob-dim"
+        >
+          <SvgIcon
+            class="mr-2"
+            icon-class="warning"
+            :svgType="SvgTypes.stroke"
+            stroke="var(--text-dim)"
+          />
+          {{ t('settings.empty-recent-comments') }}
+        </div>
       </template>
       <template v-else>
         <li
@@ -79,54 +95,126 @@
 </template>
 
 <script lang="ts">
-import { computed, defineComponent, onBeforeMount, ref, watch } from 'vue'
+import { computed, defineComponent, ref, watch } from 'vue'
 import { SubTitle } from '@/components/Title'
+import SvgIcon from '@/components/SvgIcon/index.vue'
 import { RecentComment } from '@/utils'
-import { GithubComments } from '@/utils/github-api'
-import { LeanCloudComments } from '@/utils/leancloud-api'
+import { GithubComments } from '@/utils/comments/github-api'
+import { LeanCloudComments } from '@/utils/comments/leancloud-api'
 import { useAppStore } from '@/stores/app'
 import { useI18n } from 'vue-i18n'
+import { TwikooComments } from '@/utils/comments/twikoo-api'
+import { WalineComments } from '@/utils/comments/waline-api'
+import { SvgTypes } from '@/components/SvgIcon/index.vue'
 
 export default defineComponent({
   name: 'ObRecentComment',
-  components: { SubTitle },
+  components: { SubTitle, SvgIcon },
   setup() {
     const appStore = useAppStore()
     const { t } = useI18n()
     let recentComments = ref<RecentComment[]>([])
+    let loading = ref<boolean>(true)
+
+    const enabledCommentPlugin = computed<string | undefined>(() => {
+      if (
+        !!appStore.themeConfig.plugins.gitalk.enable &&
+        !!appStore.themeConfig.plugins.gitalk.recentComment
+      ) {
+        return 'gitalk'
+      }
+
+      if (
+        !!appStore.themeConfig.plugins.valine.enable &&
+        !!appStore.themeConfig.plugins.valine.recentComment
+      ) {
+        return 'valine'
+      }
+
+      if (
+        !!appStore.themeConfig.plugins.twikoo.enable &&
+        !!appStore.themeConfig.plugins.twikoo.recentComment
+      ) {
+        return 'twikoo'
+      }
+
+      if (
+        !!appStore.themeConfig.plugins.waline.enable &&
+        !!appStore.themeConfig.plugins.waline.recentComment
+      ) {
+        return 'waline'
+      }
+
+      return undefined
+    })
 
     const initRecentComment = () => {
-      if (!appStore.configReady) return
-      if (
-        appStore.themeConfig.plugins.gitalk.enable &&
-        appStore.themeConfig.plugins.gitalk.recentComment
-      ) {
-        const githubComments = new GithubComments({
-          repo: appStore.themeConfig.plugins.gitalk.repo,
-          clientId: appStore.themeConfig.plugins.gitalk.clientID,
-          clientSecret: appStore.themeConfig.plugins.gitalk.clientSecret,
-          owner: appStore.themeConfig.plugins.gitalk.owner,
-          admin: appStore.themeConfig.plugins.gitalk.admin[0]
-        })
+      if (!appStore.configReady || enabledCommentPlugin.value === undefined) {
+        loading.value = false
+        return
+      }
 
-        githubComments.getComments().then(response => {
-          recentComments.value = response
-        })
-      } else if (
-        appStore.themeConfig.plugins.valine.enable &&
-        appStore.themeConfig.plugins.valine.recentComment
-      ) {
-        const leadCloudComments = new LeanCloudComments({
-          appId: appStore.themeConfig.plugins.valine.app_id,
-          appKey: appStore.themeConfig.plugins.valine.app_key,
-          avatar: appStore.themeConfig.plugins.valine.avatar,
-          admin: appStore.themeConfig.plugins.valine.admin,
-          lang: appStore.themeConfig.plugins.valine.lang
-        })
+      switch (enabledCommentPlugin.value) {
+        case 'gitalk':
+          const githubComments = new GithubComments({
+            repo: appStore.themeConfig.plugins.gitalk.repo,
+            clientId: appStore.themeConfig.plugins.gitalk.clientID,
+            clientSecret: appStore.themeConfig.plugins.gitalk.clientSecret,
+            owner: appStore.themeConfig.plugins.gitalk.owner,
+            admin: appStore.themeConfig.plugins.gitalk.admin[0]
+          })
 
-        leadCloudComments.getRecentComments(7).then(response => {
-          recentComments.value = response
-        })
+          githubComments.getComments().then(response => {
+            recentComments.value = response
+          })
+
+          break
+
+        case 'valine':
+          const leadCloudComments = new LeanCloudComments({
+            appId: appStore.themeConfig.plugins.valine.app_id,
+            appKey: appStore.themeConfig.plugins.valine.app_key,
+            avatar: appStore.themeConfig.plugins.valine.avatar,
+            admin: appStore.themeConfig.plugins.valine.admin,
+            lang: appStore.themeConfig.plugins.valine.lang
+          })
+
+          leadCloudComments.getRecentComments(7).then(res => {
+            recentComments.value = res
+            loading.value = false
+          })
+
+          break
+
+        case 'twikoo':
+          const twikooComments = new TwikooComments({
+            envId: appStore.themeConfig.plugins.twikoo.envId,
+            lang: appStore.themeConfig.plugins.twikoo.lang
+          })
+
+          twikooComments.getRecentComments(7).then(res => {
+            recentComments.value = res
+            loading.value = false
+          })
+
+          break
+
+        case 'waline':
+          const walineComments = new WalineComments({
+            serverURL:
+              'https://' + appStore.themeConfig.plugins.waline.serverURL,
+            lang: appStore.locale ?? 'en'
+          })
+
+          walineComments.getRecentComments(7).then(res => {
+            recentComments.value = res
+            loading.value = false
+          })
+
+          break
+
+        default:
+          loading.value = false
       }
     }
 
@@ -140,13 +228,21 @@ export default defineComponent({
       }
     )
 
-    onBeforeMount(initRecentComment)
-
     return {
+      SvgTypes,
+      isLoading: computed(() => loading.value),
       comments: computed(() => {
         return recentComments.value
       }),
+      isConfigReady: computed(() => appStore.configReady),
+      initRecentComment,
+      enabledCommentPlugin,
       t
+    }
+  },
+  mounted() {
+    if (this.isConfigReady) {
+      this.initRecentComment()
     }
   }
 })