+++ src/utils/index.ts
@@ -1,13 +1,13 @@
 export interface RecentComment {
   id: number
   body: string
-  node_id: number
+  node_id?: number
   html_url: string
   issue_url: string
   created_at: string
   updated_at: string
   author_association: string
-  filtered: boolean
+  filtered?: boolean
   user: {
     id: number
     login: string
@@ -15,7 +15,7 @@ export interface RecentComment {
     html_url: string
   }
   is_admin: boolean
-  cache_flag: boolean
+  cache_flag?: boolean
 }
 
 /**
@@ -76,22 +76,22 @@ export function formatTime(
   } else if (diff < 3600) {
     // Within 1 hour
     formattedTime =
-      String(Math.ceil(diff / 60)) + languages[configs.lang].minutes
+      String(Math.floor(diff / 60)) + languages[configs.lang].minutes
   } else if (diff < 3600 * 24) {
     // Within 1 day
     formattedTime =
-      String(Math.ceil(diff / 3600)) + languages[configs.lang].hours
+      String(Math.floor(diff / 3600)) + languages[configs.lang].hours
   } else if (diff < 3600 * 24 * 30) {
     // Within 1 month
     formattedTime =
-      String(Math.ceil(diff / 3600 / 24)) + languages[configs.lang].days
+      String(Math.floor(diff / 3600 / 24)) + languages[configs.lang].days
   } else if (diff < 3600 * 24 * 365) {
     // Within 1 year
     formattedTime =
-      String(Math.ceil(diff / 3600 / 24 / 30)) + languages[configs.lang].months
+      String(Math.floor(diff / 3600 / 24 / 30)) + languages[configs.lang].months
   } else {
     formattedTime =
-      String(Math.ceil(diff / 3600 / 24 / 365)) + languages[configs.lang].years
+      String(Math.floor(diff / 3600 / 24 / 365)) + languages[configs.lang].years
   }
 
   return configs.template.replace('[TIME]', formattedTime)