+++ src/views/Home.vue
@@ -78,7 +78,7 @@
       <div>
         <Sidebar>
           <Profile author="blog-author" />
-          <RecentComment v-if="recentCommentEnable" />
+          <RecentComment />
           <TagBox />
         </Sidebar>
       </div>
@@ -225,14 +225,6 @@ export default defineComponent({
           return null
         }
         return categoryStore.categories
-      }),
-      recentCommentEnable: computed(() => {
-        return (
-          (!!appStore.themeConfig.plugins.gitalk.enable &&
-            !!appStore.themeConfig.plugins.gitalk.recentComment) ||
-          (!!appStore.themeConfig.plugins.valine.enable &&
-            !!appStore.themeConfig.plugins.valine.recentComment)
-        )
       }),
       expanderClass,
       tabClass,